restaurant-backend/requirements.txt
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-decouple==3.8
dj-database-url==3.1.2
gunicorn==25.1.0
whitenoise==6.12.0
